pyproject.toml
[project]
name = "moonai"
version = "0.2.1"
description = "Cutting-edge framework for orchestrating role-playing, autonomous AI agents. By fostering collaborative intelligence, moonai empowers agents to work together seamlessly, tackling complex missions."
readme = "README.md"
requires-python = ">=3.10,<=3.13"
authors = [
    { name = "Bruno Bracaioli", email = "bruno@bracaiolitech.com" }
]
dependencies = [
    "pydantic>=2.6.1",
    "langchain>=0.3.1",
    "openai>=1.13.3",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp-proto-http>=1.22.0",
    "instructor>=1.3.3",
    "regex>=2024.9.11",
    "click>=8.1.7",
    "python-dotenv>=1.0.0",
    "appdirs>=1.4.4",
    "jsonref>=1.1.0",
    "json-repair>=0.25.2",
    "auth0-python>=4.7.1",
    "litellm>=1.44.22",
    "pyvis>=0.3.2",
    "uv>=0.4.25",
    "tomli-w>=1.1.0",
    "tomli>=2.0.2",
    "chromadb>=0.5.18",
    "pdfplumber>=0.11.4",
    "openpyxl>=3.1.5",
    "pytest>=8.0.0",
    "lancedb>=0.5.4",
    "pytube>=15.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.3",
    "selenium>=4.18.1",
    "docx2txt>=0.8",
    "docker>=7.1.0",
    "embedchain>=0.1.114",
]

[project.urls]
Homepage = "https://moonai.dev"
Documentation = "https://docs.moonai.dev"
Repository = "https://github.com/brunobracaioli/moonai"

[project.optional-dependencies]
agentops = ["agentops>=0.3.0"]
fastembed = ["fastembed>=0.4.1"]
pdfplumber = ["pdfplumber>=0.11.4"]
pandas = ["pandas>=2.2.3"]
openpyxl = ["openpyxl>=3.1.5"]
mem0 = ["mem0ai>=0.1.29"]

[tool.hatch.build]
packages = ["moonai", "moonai.moonai_tools"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.4.10",
    "mypy>=1.10.0",
    "pre-commit>=3.6.0",
    "mkdocs>=1.4.3",
    "mkdocstrings>=0.22.0",
    "mkdocstrings-python>=1.1.2",
    "mkdocs-material>=9.5.7",
    "mkdocs-material-extensions>=1.3.1",
    "pillow>=10.2.0",
    "cairosvg>=2.7.1",
    "moonai_tools",
    "pytest>=8.0.0",
    "pytest-vcr>=1.0.2",
    "python-dotenv>=1.0.0",
    "pytest-asyncio>=0.23.7",
    "pytest-subprocess>=1.5.2",
]

[project.scripts]
moonai = "moonai.cli.cli:moonai"

[tool.mypy]
ignore_missing_imports = true
disable_error_code = "import-untyped"
exclude = ["cli/templates"]

[tool.bandit]
exclude_dirs = ["src/moonai/cli/templates"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
